Gas Optimization Bounty submission #2 (#453)
* Token Stake Optimizations (#420)

* Update `Staker` and `StakingCondition` structs

* Update `Staker` struct
* Update `StakingCondition` struct
* Other minor changes to allow for updates

* Reorder/resize varaibles in Staking20Upgradeable.

- uint256 stakingTokenDecimals => uint16 stakingTokenDecimals
- uint256 rewardTokenDecimals => uint16 rewardTokenDecimals
- Move `stakersArray` below the fixed size variables.

* Make changes in Staking20 and Remove stakerArray

* Remove stakersArray from Staking20.sol

* Reorder variables in Staking20.sol for consistency

* run solint/prettier

* run yarn gas

* Updating gas report to ensure accuracy.

---------


* Re-add stakers array

* target gas report

* fix prettier / lint error

---------

+++ src/test/staking/TokenStake_EthReward.t.sol
@@ -307,7 +307,7 @@ contract TokenStakeEthRewardTest is BaseTest {
 
     function test_state_setTimeUnit() public {
         // set value and check
-        uint256 timeUnitToSet = 100;
+        uint80 timeUnitToSet = 100;
         vm.prank(deployer);
         stakeContract.setTimeUnit(timeUnitToSet);
         assertEq(timeUnitToSet, stakeContract.getTimeUnit());
@@ -513,7 +513,7 @@ contract TokenStakeEthRewardTest is BaseTest {
         stakeContract.stake(400);
 
         // set timeUnit to zero
-        uint256 newTimeUnit = 0;
+        uint80 newTimeUnit = 0;
         vm.prank(deployer);
         vm.expectRevert("time-unit can't be 0");
         stakeContract.setTimeUnit(newTimeUnit);
